convert filters to array
The fact that we're using `Object.keys(filters).some(` is pretty
telling that this should have been a list/array to begin with.

Explain the filter regex and give it a name for readability. Log
a warning if a filter is not considered valid. The regex has group-
captures in it, but they go unused. Maybe drop them?