make the filter function readable
Working from the previous fix where we turned a hash of filters into
a list, we now attempt to make the code using said filters more
readable.

It took some time to understand that we weren't actually using the
some/every methods for their original purpose, but rather as an
optimization to break out of a forEach-loop.

It could probably be rewritten to use the `every(` method and its
return-value to say wether or not an element is allowed.