use array filters
Create wrappers and use the built-in .filter( method in JS-arrays,
and chain them together to apply the filter function we previously
rewrote to work as a proper filter callback function.

Splitting the chained functions over multiple lines seems a common
practice in the world of JS, and seems to work with duktape.

someOject.someArray
    .filter(onThing)
    .filter(onOther)
    .forEach(...

instead of

someOject.someArray.filter(onThing).filter(onOther).forEach(...